use CoverTarget for Subscriptions, some cleanup of how it looks
