+++ app/src/main/res/layout/subscription_item.xml
@@ -18,16 +18,11 @@
     <TextView
         android:id="@+id/txtvTitle"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:layout_margin="@dimen/widget_margin"
+        android:layout_height="match_parent"
         android:ellipsize="end"
-        android:padding="@dimen/widget_margin"
         style="@style/AntennaPod.TextView.Heading"
-        android:textSize="15sp"
         android:textStyle="bold"
-        android:layout_gravity="bottom"
-        android:textColor="@android:color/white"
+        android:gravity="center_horizontal|bottom"
         android:background="#55000000"
         tools:text="@string/app_name" />
